+++ NEWS.md
@@ -1,6 +1,8 @@
 # cffr (development version)
 
 - Add DOI <https://doi.org/10.5281/zenodo.5509766>
+- Citation of installed packages extracted using `citation()`
+- Auto-generating `preferred-citation` key from DESCRIPTION.
 
 # cffr 0.0.1